Rethinking the API.
Now it'll focus on creating objects that are meaningful in an
multimedia synthesis context: sounds, images, 3D models and so on.

I'm keeping types like `Transaction`, `Operator`, `Function` and so on
out of the public API because you really just need to think in terms
of code, e.g. strings. I'm making an analogy to the way SQL APIs
generally work: you write SQL in a string, and you don't worry about
the internal data structures it uses to parse that.